Add:キーワード検索機能を追加しました。店名とキャッチフレーズとフクレココメントに対して部分一致検索します
